Add generic BL31 parameter parsing code
This patch adds code that can be used by a platform implementation to
parse a list of platform parameters passed in by the previous boot stage
(BL2). It is based of a format that was originally designed for the
Rockchip platform (and is binary backwards-compatible to it). This patch
makes this mechanism available to be reused on other platforms and also
opens the possibility of supporting platform-independent parameters that
can be handled by common code.

Change-Id: I57b14145725f4b48ff0e3192990c78d5aa5b4acb
Signed-off-by: Julius Werner <jwerner@chromium.org>