coreboot: Add support for CBMEM console
coreboot supports an in-memory console to store firmware logs even when
no serial console is available. It is widely supported by
coreboot-compatible bootloaders (including SeaBIOS and GRUB) and can be
read by the Linux kernel.

This patch allows BL31 to add its own log messages to this console. Its
presence can be detected automatically through the coreboot table if
coreboot support is compiled in. By hooking directly into putchar(), it
can work side-by-side with the existing serial console, and is
unaffected by the usual console_uninit() at the end of EL3 runtime init
(allowing it to continue logging messages that are generated by runtime
services).

Change-Id: I31254dfa0c2fdeb7454634134b5707b4b4154907
Signed-off-by: Julius Werner <jwerner@chromium.org>