Allow coreboot CBMEM console to log crash dumps
This patch hooks up the coreboot CBMEM console to the crash reporting
infrastructure. Like for normal log messages, this is implemented as an
optional addition that can run beside the normal (serial) console.

Since the crash reporting infrastructure was written with only a single,
platform-specific console backend in mind, this patch wraps all calls to
plat_crash_console_putc/_flush in a macro that will call all compiled-in
crash console backends in succession. Each backend still needs to comply
with the tight ABI restrictions set out for each of those functions
(including preservation of parameters for the next backend).

Since the CBMEM console is an in-memory console intended to survive
reboots, the "flush" primitive corresponds to a cache flush of the
console buffer. Unfortunately, it's not really possible to squeeze a
cache flush routine in the only two registers allowed by the
plat_crash_console_flush ABI. Therefore, this patch changes the ABI to
allow four registers (x0 - x3, which doesn't pose a problem at any of
the current call sites of the function).

All changes in this patch only affect the AArch64 version of the crash
reporting infrastructure, since coreboot only supports Trusted Firmware
integration for AArch64. Equivalent changes could be made to the AArch32
versions of respective files at a later point if the need arises.

Change-Id: I0a2037944d3ddd77c052638e09a1b01ad1e7fed1
Signed-off-by: Julius Werner <jwerner@chromium.org>